Adding the DragandDrop functionality to the feature branch (#62)
* Added new Drag Functionality

* Finalised Drag Functionality with comments explaining the code

* removed extra call to playMusic() aas per comment in PR #62

* Revert "removed extra call to playMusic() as per comment in PR #62"

This reverts commit 692006a292329eebe39379801dca812195a812a6.

* fixed double call to playMusic in LevelGameArea

---------
