+++ source/core/src/main/com/csse3200/game/areas/LevelGameArea.java
@@ -9,9 +9,9 @@ import com.csse3200.game.entities.factories.GridFactory;
 import com.csse3200.game.services.ResourceService;
 import com.csse3200.game.services.ServiceLocator;
 import com.csse3200.game.components.gamearea.GameAreaDisplay;
-import com.csse3200.game.ui.DragAndDropDemo;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+
 
 public class LevelGameArea extends GameArea{
     private static final Logger logger = LoggerFactory.getLogger(LevelGameArea.class);
@@ -51,56 +51,10 @@ public class LevelGameArea extends GameArea{
         displayUI();
 
         spawnMap();
+
         float scale = 1.4f;
         spawnTiles(scale);
-        testUI_1();
-
-        testUI_2();
-        testUI_3();
-
         playMusic();
-
-    }
-
-    private void testUI_1() {
-//    Entity dragUI = new Entity();
-//    dragUI.addComponent(new com.csse3200.game.ui.DragAndDropDemo());
-//    spawnEntity(dragUI);
-        Entity ui = new Entity();
-        DragAndDropDemo dragUI = new DragAndDropDemo();
-        dragUI.setTexture("images/ghost_1.png"); // choose which image
-        dragUI.setOffsets(0f, 500f); // bottom-right-ish
-        dragUI.setScale(0.1f); // smaller
-        ui.addComponent(dragUI);
-        spawnEntity(ui);
-
-    }
-
-    private void testUI_2() {
-//    Entity dragUI = new Entity();
-//    dragUI.addComponent(new com.csse3200.game.ui.DragAndDropDemo());
-//    spawnEntity(dragUI);
-        Entity ui = new Entity();
-        DragAndDropDemo dragUI = new DragAndDropDemo();
-        dragUI.setTexture("images/box_boy.png"); // choose which image
-        dragUI.setOffsets(0f, 0f); // bottom-right-ish
-        dragUI.setScale(0.1f); // smaller
-        ui.addComponent(dragUI);
-        spawnEntity(ui);
-
-    }
-
-    private void testUI_3() {
-//    Entity dragUI = new Entity();
-//    dragUI.addComponent(new com.csse3200.game.ui.DragAndDropDemo());
-//    spawnEntity(dragUI);
-        Entity ui = new Entity();
-        DragAndDropDemo dragUI = new DragAndDropDemo();
-        dragUI.setTexture("images/ghost_1.png"); // choose which image
-        dragUI.setOffsets(0f, 250f); // bottom-right-ish
-        dragUI.setScale(0.1f); // smaller
-        ui.addComponent(dragUI);
-        spawnEntity(ui);
 
     }
